refactor(tui): split order/tree/app by responsibility, trim comments, full asserts
Responds to review feedback on PR #112 (ADR 0035, tests-sorted-last):

- Split rinkaku-tui/src/order.rs into order/{mod,rank,sort}.rs: rank.rs
  holds directory ranking + cycle analysis (DirCondensation,
  rank_directories, cycle_partners/edges/explanation), sort.rs holds
  sibling ordering (OrderMode, order_tree, SiblingTier). mod.rs is pure
  re-export dispatch. order_tests/ split the same way into
  order/{rank_tests,sort_tests}/.
- Split rinkaku-tui/src/tree.rs into tree/{mod.rs,tests_section.rs}:
  the ADR 0035 Phase B "whole test file" detection and Section-wrapping
  logic (is_whole_test_file, wrap_section, sort_alphabetically) moves
  to its own module with its own unit tests
  (tree/tests_section_tests/), leaving tree/mod.rs to the general tree-
  building machinery. tree_tests/ moves alongside as tree/tree_tests/.
- Split rinkaku-tui/src/app/mod.rs's key-event dispatch
  (handle_key/handle_key_with_popup_open/handle_scroll_key, ~700
  lines) into app/handle_key.rs as a second `impl App` block, resolving
  a pre-existing file-size-warning threshold breach (1504 -> 1527 ->
  805 lines) the map-assisted review pass flagged as a review-blocker.
- Trimmed restating doc/inline comments across the new Section-
  handling code (tree::NodeKind::Section, TESTS_SECTION_PATH,
  row_view's Section/test-badge arms, app's Section match arms) to
  keep only WHY/constraint content per CLAUDE.md's comment discipline
  (also just codified in #113): why a distinct NodeKind variant instead
  of a synthetic Dir path, why blast-radius/diff-target/detail return
  NotApplicable/None for a Section row, why `label` goes unused.
- Rewrote 5 of 6 tree_tests/tests_section.rs tests from field-by-field
  partial asserts to whole-Tree/TreeNode pretty_assertions::assert_eq!
  comparisons, per the repo's fully-qualified-assert convention.
- ADR 0035: Status proposed -> accepted.

make test: 1121 tests passing (169 rinkaku + 361 rinkaku-core + 591
rinkaku-tui). make lint: clean.
