+++ rinkaku-tui/src/order/rank_tests/cycle_partners.rs
