+++ rinkaku-tui/src/order/rank_tests/scc_helpers.rs
