+++ rinkaku-tui/src/tree/tree_tests/symbol_badges.rs
